+++ kwimage/__main__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+# PYTHON_ARGCOMPLETE_OK
+if __name__ == '__main__':
+    from kwimage.cli.__main__ import main
+    main()